BREAKING CHANGES: add much to the project!
this commit ships with a series of modifications to the config files.
includes: kanta support (to remapping caps-lock onto esc, and when
holds go to ctrl), tailscale support, nvim/emacs configured when
upgrading the system, and remove sudo!

please, take note that this is a experimental commit, and may break
your system completly! so, update wisely, and if you got some error,
feel free to post a issue in github